US EPA AQI Index levels, names, colors and action text

Official, US EPA AQI levels, names, colors and action text are provided in a list for easy coloring and labeling.

US_AQI

Format

A list with named elements

Details

AQI breaks and associated names and colors

Note

The low end of each break category is used as the breakpoint.

Breaks

Breakpoints are given in units reported for each parameter and include:

breaks_AQI

breaks_CO

breaks_NO2

breaks_OZONE_1hr

breaks_OZONE_8hr

breaks_PM2.5

breaks_PM10

Colors

Several different color palettes are provided:

colors_EPA -- official EPA AQI colors

colors_subdued -- subdued colors fo use with leaflet maps

colors_deuteranopia -- color vision impaired colors

Names

Names of AQI categories are provided in several languages identified by the ISO 639-2 alpha-3 code:

names_eng

names_spa

Actions

Text for "actions to protect yourself" are provided for each category in several languages identified by the ISO 639-2 alpha-3 code:

actions_eng

actions_spa

Currently supported languages include English (eng) and Spanish (spa).

AQI breaks are defined at https://aqs.epa.gov/aqsweb/documents/codetables/aqi_breakpoints.html and are given in units appropriate for each pollutant.

AQI colors are defined at https://docs.airnowapi.org/aq101

Examples

print(US_AQI$breaks_AQI)

#> [1] -Inf   50  100  150  200  300  Inf

print(US_AQI$colors_EPA)

#> [1] "#00E400" "#FFFF00" "#FF7E00" "#FF0000" "#8F3F97" "#7E0023"

print(US_AQI$names_eng)

#> [1] "Good"           "Moderate"       "USG"            "Unhealthy"     
#> [5] "Very Unhealthy" "Hazardous"     

print(US_AQI$names_spa)

#> [1] "Buena"         "Moderada"      "IGS"           "Insalubre"    
#> [5] "Muy insalubre" "Peligrosa"