Load most recent AirNow monitoring data

Loads pre-generated .rda files containing the most recent AirNow data.

If archiveDataDir is defined, data will be loaded from this local archive. Otherwise, data will be loaded from the monitoring data repository maintained by the USFS AirFire team.

The files loaded by this function are updated multiple times an hour and contain data for the previous 10 days.

For daily updates covering the most recent 45 days, use airnow_loadDaily().

For data extended more than 45 days into the past, use airnow_loadAnnual().

Pre-processed AirNow exists for the following parameters:

PM2.5

PM2.5_nowcast

airnow_loadLatest(
  archiveBaseUrl = paste0("https://airfire-data-exports.s3.us-west-2.amazonaws.com/",
    "monitoring/v2"),
  archiveBaseDir = NULL,
  QC_negativeValues = c("zero", "na", "ignore"),
  parameterName = "PM2.5"
)

Arguments

archiveBaseUrl

Base URL for monitoring v2 data files.

archiveBaseDir

Local base directory for monitoring v2 data files.

QC_negativeValues

Type of QC to apply to negative values.

parameterName

One of the EPA AQS criteria parameter names.

Value

A mts_monitor object with AirNow data. (A list with meta and data dataframes.)

See also

Examples