Load daily AIRSIS monitoring data

Loads pre-generated .rda files containing daily AIRSIS data.

If archiveDataDir is defined, data will be loaded from this local archive. Otherwise, data will be loaded from the monitoring data repository maintained by the USFS AirFire team.

The files loaded by this function are updated once per day and contain data for the previous 45 days.

For the most recent data in the last 10 days, use airsis_loadLatest().

For data extended more than 45 days into the past, use airsis_loadAnnual().

airsis_loadDaily(
  archiveBaseUrl = paste0("https://airfire-data-exports.s3.us-west-2.amazonaws.com/",
    "monitoring/v2"),
  archiveBaseDir = NULL,
  QC_negativeValues = c("zero", "na", "ignore"),
  QC_removeSuspectData = TRUE
)

Arguments

archiveBaseUrl

Base URL for monitoring v2 data files.

archiveBaseDir

Local base directory for monitoring v2 data files.

QC_negativeValues

Type of QC to apply to negative values.

QC_removeSuspectData

Removes monitors determined to be misbehaving.

Value

A mts_monitor object with AIRSIS data. (A list with meta and data dataframes.)

Note

Some older AIRSIS timeseries contain only values of 0, 1000, 2000, 3000, ... ug/m3. Data from these deployments pass instrument-level QC checks but these timeseries generally do not represent valid data and should be removed. With QC_removeSuspectData = TRUE (the default), data is checked and periods reporting only values of 0:10 * 1000 ug/m3 are invalidated.

Only those personally familiar with the individual instrument deployments should work with the "suspect" data.

See also

Examples