Load annual Clarity monitoring data

Loads pre-generated .rda files containing annual Clarity data.

If archiveDataDir is defined, data will be loaded from this local archive. Otherwise, data will be loaded from the monitoring data repository maintained by the USFS AirFire team.

Current year files loaded by this function are updated once per week.

For the most recent data in the last 10 days, use clarity_loadLatest().

clarity_loadAnnual(
  year = NULL,
  archiveBaseUrl = paste0("INSERT_BASE_HERE", "sensors/v3/PM2.5"),
  archiveBaseDir = NULL,
  QC_negativeValues = c("zero", "na", "ignore"),
  QC_calibration = c("global PM2.5 v2.1", "all")
)

Arguments

year

Year [YYYY].

archiveBaseUrl

Base URL for monitoring v2 data files.

archiveBaseDir

Local base directory for monitoring v2 data files.

QC_negativeValues

Type of QC to apply to negative values.

QC_calibration

Clairty calibration category

Value

A mts_monitor object with Clarity data. (A list with meta and data dataframes.)

Note

The QC_calibration argument specifies the calibration/correction algorithm used by Clarity for PM2.5 measurements. See Clarity v2 clibration model It is recommended that users stick with the default calibration: "global PM2.5 v2.1". Specifying "all" will retrieve measurements harvested before the introduction of "global PM2.5 v2.1" and are not directly comparable with more recent measurements.

Disclaimer

Data archives accessed by this function are not maintained by Clarity and are provided with no guarantee of completeness or future maintenance. They may be removed at any time with no warning.

See also

Examples

if (FALSE) { # \dontrun{
library(AirMonitor)

# Fail gracefully if any resources are not available
try({

# 2025 Emigrant Fire
# See https://inciweb.wildfire.gov/incident-information/orwif-emigrant-fire

# Clarity monitors in Oak Ridge, OR (-122.4670, 43.7473)
clarity_loadAnnual(2025) %>%
  monitor_filterByDistance(-122.4670, 43.7473, 10000) %>%
  monitor_collapse(deviceID = "Oak_Ridge") %>%
  monitor_filterDate(20250815, 20250915, timezone = "America/Los_Angeles") %>%
  monitor_dailyBarplot(main = "Clarity Monitors in Oak Ridge, Oregon")

}, silent = FALSE)
} # }