Names of standard metadata columns

Vector of names of the required monitor$meta columns. These represent metadata columns that must exist in every valid mts_monitor object. Any number of additional columns may also be present.

coreMetadataNames

Format

A vector of character strings

Details

coreMetadataNames

Examples

 print(coreMetadataNames, width = 80)
#>  [1] "deviceDeploymentID"    "deviceID"              "deviceType"           
#>  [4] "deviceDescription"     "deviceExtra"           "pollutant"            
#>  [7] "units"                 "dataIngestSource"      "dataIngestURL"        
#> [10] "dataIngestUnitID"      "dataIngestExtra"       "dataIngestDescription"
#> [13] "locationID"            "locationName"          "longitude"            
#> [16] "latitude"              "elevation"             "countryCode"          
#> [19] "stateCode"             "countyName"            "timezone"             
#> [22] "houseNumber"           "street"                "city"                 
#> [25] "zip"                   "AQSID"                 "fullAQSID"