Collapse an mts_monitor object into a single time series

Collapses data from all time series in mts_monitor into a single-time series mts_monitor object using the function provided in the FUN argument. The single-time series result will be located at the mean longitude and latitude unless longitude and latitude parameters are specified.

Any columns of meta that are constant across all records will be retained in the returned mts_monitor meta.

The core metadata associated with this location (e.g. countryCode, stateCode, timezone, ...) will be determined from the most common (or average) value found in mts$meta. This will be a reasonable assumption for the vast majority of intended use cases where data from multiple instruments in close proximity are averaged together.

monitor_collapse(
  monitor,
  longitude = NULL,
  latitude = NULL,
  deviceID = "generatedID",
  FUN = mean,
  na.rm = TRUE,
  ...
)

Arguments

monitor
mts_monitor object.

longitude
Longitude of the collapsed time series.

latitude
Latitude of the collapsed time series.

deviceID
Device identifier for the collapsed time series.

FUN
Function used to collapse multiple time series.

na.rm
Logical specifying whether NA values should be ignored when FUN is applied.

...
additional arguments to be passed on to the apply() function.

Value

A mts_monitor object representingn a single time series

Note

After FUN is applied, values of +/-Inf and NaN are converted to NA. This is a convenience for the common case where FUN = min/max or FUN = mean and some of the time steps have all missing values. See the R documentation for min for an explanation.