Calculate distances from mts_monitor locations to a location of interest

This function returns the distances (meters) between monitor locations and a location of interest. These distances can be used to create a mask identifying monitors within a certain radius of the location of interest.

monitor_distance(
  monitor = NULL,
  longitude = NULL,
  latitude = NULL,
  measure = c("geodesic", "haversine", "vincenty", "cheap")
)

Arguments

monitor
mts_monitor object.

longitude
Longitude of the location of interest.

latitude
Latitude of the location of interest.

measure
One of "geodesic", "haversine" "vincenty", or "cheap".

Value

Vector of of distances (meters) named by deviceDeploymentID.

Note

The measure "cheap" may be used to speed things up depending on the spatial scale being considered. Distances calculated with measure = "cheap" will vary by a few meters compared with those calculated using measure = "geodesic".

Examples

library(AirMonitor)

PacNW <- example_88101

# Walla Walla
lon <- -118.330278
lat <- 46.065

distance <- monitor_distance(PacNW, lon, lat)

#> Warning: Unknown or uninitialised column: `monitorID`.

closestIndex <- which(distance == min(distance))

print("Distances from Walla Walla")

#> [1] "Distances from Walla Walla"

sprintf(
  "%s: %d km",
  PacNW$meta$locationName,
  round(distance / 1000)
)

#>  [1] "Ballard Road: 579 km"                                 
#>  [2] "NA: 689 km"                                           
#>  [3] "Minidoka National Historic Site: 497 km"              
#>  [4] "Salmon - Charles St.: 360 km"                         
#>  [5] "Pinehurst: 229 km"                                    
#>  [6] "NA: 53 km"                                            
#>  [7] "Wenatchee-Fifth St: 216 km"                           
#>  [8] "Vancouver-Ne 84th Ave: 334 km"                        
#>  [9] "Vancouver-Ne 84th Ave: 334 km"                        
#> [10] "Seattle-10th & Weller: 349 km"                        
#> [11] "Seattle - Duwamish: 348 km"                           
#> [12] "Seattle - Beacon Hill: 347 km"                        
#> [13] "Kent - James & Central: 332 km"                       
#> [14] "Bremerton-Spruce Ave: 369 km"                         
#> [15] "Ellensburg - Ruby St: 199 km"                         
#> [16] "Tacoma - L Street: 339 km"                            
#> [17] "Anacortes-202 Ave: 424 km"                            
#> [18] "Lynnwood - 212th: 360 km"                             
#> [19] "Lynnwood - 212th: 360 km"                             
#> [20] "Darrington - Fir St (Darrington High School): 347 km" 
#> [21] "Marysville - 7th Ave (Marysville Junior High): 366 km"
#> [22] "Marysville - 7th Ave (Marysville Junior High): 366 km"
#> [23] "Spokane - Augusta Ave: 193 km"                        
#> [24] "Spokane - Augusta Ave: 193 km"                        
#> [25] "Bellingham - Yew Street: 431 km"                      
#> [26] "Yakima - 4th Ave: 177 km"                             
#> [27] "Yakima - 4th Ave: 177 km"