Data-based subsetting of time series within an mts_monitor object.

Subsetting of monitor acts similarly to tidyselect::where() working on monitor$data. The returned mts_monitor object will contain only those time series where FUN applied to the time series data returns TRUE.

monitor_selectWhere(monitor, FUN)

Arguments

monitor
mts_monitor object.

FUN
A function applied to time series data that returns TRUE or FALSE.

Value

A subset of the incoming mts_monitor object. (A list with meta and data dataframes.)

See also

Examples

library(AirMonitor)

# Show all Camp_Fire locations
Camp_Fire$meta$locationName

#>   [1] "Livermore - Rincon"                     
#>   [2] "Oakland"                                
#>   [3] "Oakland West"                           
#>   [4] "Laney College"                          
#>   [5] "Berkeley Aquatic Park"                  
#>   [6] "Pleasanton - Owens Ct."                 
#>   [7] "Chico -  East"                          
#>   [8] "Gridley"                                
#>   [9] "San Andreas"                            
#>  [10] "Cortina Rancheria"                      
#>  [11] "Colusa - Sunrise Blvd."                 
#>  [12] "Concord"                                
#>  [13] "San Pablo - Rumrill"                    
#>  [14] "Fresno - Garland"                       
#>  [15] "Huron"                                  
#>  [16] "Tranquillity"                           
#>  [17] "Willows-Colusa"                         
#>  [18] "Calexico - Ethel Street"                
#>  [19] "El Centro - 9th Street"                 
#>  [20] "NCORE - Bishop"                         
#>  [21] "Keeler"                                 
#>  [22] "Bishop Paiute Tribe-CA"                 
#>  [23] "Mojave - 923 Poole St."                 
#>  [24] "Bakersfield - California Ave"           
#>  [25] "Lebec"                                  
#>  [26] "Glendora - Laurel"                      
#>  [27] "Los Angeles - N. Main Street"           
#>  [28] "Reseda"                                 
#>  [29] "South Long Beach"                       
#>  [30] "Santa Clarita"                          
#>  [31] "Lancaster-Division"                     
#>  [32] "Madera-City"                            
#>  [33] "San Rafael"                             
#>  [34] "Yosemite Village - Visitor Center"      
#>  [35] "Ukiah Library"                          
#>  [36] "Willits Unified School District"        
#>  [37] "Merced - S. Coffee Ave"                 
#>  [38] "Mammoth Lakes"                          
#>  [39] "Lee Vining"                             
#>  [40] "Carmel Valley AMS"                      
#>  [41] "King City AMS"                          
#>  [42] "Salinas AMS"                            
#>  [43] "Napa - Napa Valley College"             
#>  [44] "Grass Valley"                           
#>  [45] "Truckee BAM"                            
#>  [46] "Anaheim"                                
#>  [47] "Auburn"                                 
#>  [48] "Colfax"                                 
#>  [49] "Tahoe City"                             
#>  [50] "Lincoln"                                
#>  [51] "Chester BAM"                            
#>  [52] "Portola"                                
#>  [53] "Quincy BAM"                             
#>  [54] "Joshua Tree NP - Cottonwood Canyon 2"   
#>  [55] "Banning - South Hathaway Street"        
#>  [56] "Temecula (Lake Skinner)"                
#>  [57] "Riverside - Rubidoux"                   
#>  [58] "Mira Loma - Van Buren"                  
#>  [59] "Lake Elsinore - W. Flint Street"        
#>  [60] "Arden Arcade - Del Paso Manor"          
#>  [61] "Downtown Sacramento - T Street"         
#>  [62] "Elk Grove"                              
#>  [63] "Folsom"                                 
#>  [64] "Sloughhouse"                            
#>  [65] "Hollister AMS"                          
#>  [66] "Crestline - Lake Gregory"               
#>  [67] "Victorville - Park Avenue"              
#>  [68] "Upland"                                 
#>  [69] "Alpine"                                 
#>  [70] "Camp Pendleton"                         
#>  [71] "Otay Mesa - Donovan"                    
#>  [72] "El Cajon - Lexington Elementary School" 
#>  [73] "San Francisco"                          
#>  [74] "Stockton - Hazelton Street"             
#>  [75] "TracyAP"                                
#>  [76] "Mesa 2 /Nipomo"                         
#>  [77] "San Luis Obispo - Higuera St"           
#>  [78] "Nipomo Mesa CDF"                        
#>  [79] "Atascadero2"                            
#>  [80] "Redwood City"                           
#>  [81] "Santa Barbara"                          
#>  [82] "Santa Maria - Broadway"                 
#>  [83] "Lompoc H Street"                        
#>  [84] "Goleta"                                 
#>  [85] "Gilory - 9th Street"                    
#>  [86] "San Jose - Jackson St."                 
#>  [87] "San Jose - Knox Ave"                    
#>  [88] "Santa Cruz AMS"                         
#>  [89] "Zayante Fire Station"                   
#>  [90] "Felton Cal-Fire"                        
#>  [91] "Boulder Creek"                          
#>  [92] "Anderson"                               
#>  [93] "Yreka"                                  
#>  [94] "Vallejo"                                
#>  [95] "Vacaville"                              
#>  [96] "Rio Vista"                              
#>  [97] "Sebastopol"                             
#>  [98] "Turlock - S. Minaret Street"            
#>  [99] "Yuba City"                              
#> [100] "Red Bluff - Walnut office"              
#> [101] "Weaverville"                            
#> [102] "Sequoia and Kings Canyon - Ash Mountain"
#> [103] "Visalia - N. Church Street"             
#> [104] "Porterville-NFS"                        
#> [105] "Thousand Oaks - Moorpark Road"          
#> [106] "Piru - Pacific"                         
#> [107] "Ojai - Ojai Ave."                       
#> [108] "Simi Valley - Cochran Street"           
#> [109] "El Rio - Rio Mesa School #2"            
#> [110] "Davis - UCD Campus"                     
#> [111] "Woodland"                               
#> [112] "EBAM7 - Sac 5th St. Warehouse"          
#> [113] "Pinehurst"                              
#> [114] "Kernville"                              
#> [115] "Kennedy Meadows"                        
#> [116] "Modesto - 14th Street"                  
#> [117] "Lone Pine Paiute/Shoshone Tribe-CA"     
#> [118] "SLV Middle School AMS"                  
#> [119] "Roseville - N.Sunrise/Douglas"          
#> [120] "Paradise Theater"                       
#> [121] "ARB 1017 (offline)"                     
#> [122] "LCAQMD1001"                             
#> [123] "NPT1002"                                
#> [124] "NPT1002"                                
#> [125] "1051 Kernville EBAM"                    
#> [126] "1060 North Fork"                        
#> [127] "ARB 1041 (offline)"                     
#> [128] "ARB 1011 (Trinity C)"                   
#> [129] "Mariposa 1000"                          
#> [130] "SEKI1032 EBAM4"                         
#> [131] "ARB 1025 (offline)"                     
#> [132] "ARB 1023 (offline)"                     
#> [133] "ARB 1010 (offline)"                     

# Use package US_AQI data for HAZARDOUS
name <- US_AQI$names_eng[6]
threshold <- US_AQI$breaks_PM2.5[6]

# Find HAZARDOUS locations
worst_sites <-
  Camp_Fire %>%
  monitor_selectWhere(
    function(x) { any(x >= threshold, na.rm = TRUE) }
  )

# Show the worst locations
worst_sites$meta$locationName

#>  [1] "Livermore - Rincon"             "Pleasanton - Owens Ct."        
#>  [3] "Chico -  East"                  "Gridley"                       
#>  [5] "Colusa - Sunrise Blvd."         "San Pablo - Rumrill"           
#>  [7] "Willows-Colusa"                 "Ukiah Library"                 
#>  [9] "Salinas AMS"                    "Grass Valley"                  
#> [11] "Colfax"                         "Lincoln"                       
#> [13] "Arden Arcade - Del Paso Manor"  "Downtown Sacramento - T Street"
#> [15] "Elk Grove"                      "Folsom"                        
#> [17] "Sloughhouse"                    "Stockton - Hazelton Street"    
#> [19] "TracyAP"                        "Anderson"                      
#> [21] "Rio Vista"                      "Sebastopol"                    
#> [23] "Turlock - S. Minaret Street"    "Yuba City"                     
#> [25] "El Rio - Rio Mesa School #2"    "Davis - UCD Campus"            
#> [27] "Woodland"                       "Modesto - 14th Street"         
#> [29] "Roseville - N.Sunrise/Douglas"  "1060 North Fork"               
#> [31] "ARB 1041 (offline)"             "ARB 1023 (offline)"            
#> [33] "ARB 1010 (offline)"