Load annual EPA AQS monitoring data

Loads pre-generated .RData files containing annual EPA data.

EPA parameter codes include:

44201 -- Ozone

42101 -- CO

88101 -- PM2.5 FRM/FEM Mass (begins in 2008)

88502 -- PM2.5 non FRM/FEM Mass (begins in 1998)

81102 -- PM10 Mass (begins in 1988)

epa_aqs_loadAnnual(
  parameterCode = NULL,
  year = NULL,
  archiveBaseUrl = NULL,
  archiveBaseDir = NULL,
  QC_negativeValues = c("zero", "na", "ignore")
)

Arguments

parameterCode
EPA AQS Parameter Code.

year
Desired year (integer or character representing YYYY).

archiveBaseUrl
Base URL for annual EPA AQS data files.

archiveBaseDir
Local base directory for annual EPA AQS data files.

QC_negativeValues
Type of QC to apply to negative values.

Value

A mts_monitor object with EPA AQS data.

QC

EPA AQS hourly files are highly vetted but may still have some problems. One particular issue is the presence of aphysical, negative values. (There is no such thing as a negative concentration.) These are most likely the result of instrument drift and could potentially be carefully adjusted if additional, enegineering-level raw data were available.

But in the context of large collections of aggregated data, we must take a coarser appraoch. Three types of corrections are available through the QC_negativeValues parameter:

"zero" -- replace negative values with 0

"na" -- replace negative values with NA

"ignore" -- do not replace negative values