Combine multiple mts_monitor objects

Create a combined mts_monitor from any number of mts_monitor objects or from a list of mts_monitor objects. The resulting mts_monitor object with contain all deviceDeploymentIDs found in any incoming mts_monitor and will have a regular time axis covering the the entire range of incoming data.

If incoming time ranges are tempporally non-contiguous, the resulting mts_monitor will have gaps filled with NA values.

An error is generated if the incoming mts_monitor objects have non-identical metadata for the same deviceDeploymentID unless replaceMeta = TRUE.

monitor_combine(..., replaceMeta = FALSE)

Arguments

...
Any number of valid emphmts_monitor objects or a list of objects.

replaceMeta
Logical specifying whether to allow replacement of metadata associated with deviceDeploymentIDs.

Value

A combined mts_monitor object. (A list with meta and data dataframes.)

Note

Data are combined with a "latest is best" sensibility where any data overlaps exist. Incoming mts_monitor objects are ordered based on the time stamp of their last record. Any data records found in a "later" mts_monitor will overwrite data associated with an "earlier" mts_monitor.

Examples

library(AirMonitor)

# Washington State University
Pullman <-
  NW_Megafires %>%
  monitor_select("089a067f92712ad1_530750003") %>%
  monitor_filterDatetime(2015080118, 2015080203)

# University of Idaho
Moscow <-
  NW_Megafires %>%
  monitor_select("d121a99bc6c2ac7f_160570005") %>%
  monitor_filterDatetime(2015080200, 2015080206)

monitor_combine(Pullman, Moscow) %>%
  monitor_getData()

#>               datetime d121a99bc6c2ac7f_160570005 089a067f92712ad1_530750003
#> 1  2015-08-02 01:00:00                         NA                       25.8
#> 2  2015-08-02 02:00:00                         NA                       45.4
#> 3  2015-08-02 03:00:00                         NA                       46.6
#> 4  2015-08-02 04:00:00                         NA                       45.2
#> 5  2015-08-02 05:00:00                         NA                       45.8
#> 6  2015-08-02 06:00:00                         NA                       44.9
#> 7  2015-08-02 07:00:00                       37.5                       43.9
#> 8  2015-08-02 08:00:00                       35.3                       43.7
#> 9  2015-08-02 09:00:00                       28.8                       42.2
#> 10 2015-08-02 10:00:00                       26.9                         NA
#> 11 2015-08-02 11:00:00                       26.6                         NA
#> 12 2015-08-02 12:00:00                       26.5                         NA