Daily counts of values at or above a threshold

Calculates the number of hours per day each monitor in monitor was at or above a given threshold.

Because the returned mts_monitor object is defined on a daily axis in a specific time zone, it is important that the incoming monitor contain timeseries associated with a single time zone.

monitor_dailyThreshold(
  monitor = NULL,
  threshold = NULL,
  na.rm = TRUE,
  minHours = 18,
  dayBoundary = c("clock", "LST")
)

Arguments

monitor
mts_monitor object.

threshold
AQI level name (e.g. "unhealthy") or numerical threshold at or above which a measurement is counted.

na.rm
Logical value indicating whether NA values should be ignored.

minHours
Minimum number of valid hourly records per day required to calculate statistics. Days with fewer valid records will be assigned NA.

dayBoundary
Treatment of daylight savings time: "clock" uses daylight savings time as defined in the local timezone, "LST" uses "local standard time" all year round.

Value

A mts_monitor object containing daily counts of hours at or above a threshold value..

Note

When dayBoundary = "clock", the returned monitor$data$datetime time axis will be defined in the local timezone (not "UTC") with days defined by midnight as it appears on a clock in that timezone. The transition from DST to standard time will result in a 23 hour day and standard to DST in a 25 hour day.

When dayBoundary = "LST", the returned monitor$data$datetime time axis will be defined in "UTC" with times as they appear in standard time in the local timezone. These days will be one hour off from clock time during DST but every day will consist of 24 hours.

Examples

Carmel_Valley %>%
  monitor_dailyThreshold("Moderate") %>%
  monitor_extractData()

#> # A tibble: 24 × 2
#>    datetime            a9572a904a4ed46d_060530002_03
#>    <dttm>                                      <int>
#>  1 2016-07-22 00:00:00                             0
#>  2 2016-07-23 00:00:00                             3
#>  3 2016-07-24 00:00:00                            11
#>  4 2016-07-25 00:00:00                            16
#>  5 2016-07-26 00:00:00                            19
#>  6 2016-07-27 00:00:00                             8
#>  7 2016-07-28 00:00:00                             6
#>  8 2016-07-29 00:00:00                             4
#>  9 2016-07-30 00:00:00                             8
#> 10 2016-07-31 00:00:00                            17
#> # … with 14 more rows

Carmel_Valley %>%
  monitor_dailyThreshold("Unhealthy") %>%
  monitor_extractData()

#> # A tibble: 24 × 2
#>    datetime            a9572a904a4ed46d_060530002_03
#>    <dttm>                                      <int>
#>  1 2016-07-22 00:00:00                             0
#>  2 2016-07-23 00:00:00                             0
#>  3 2016-07-24 00:00:00                             0
#>  4 2016-07-25 00:00:00                             0
#>  5 2016-07-26 00:00:00                            14
#>  6 2016-07-27 00:00:00                             1
#>  7 2016-07-28 00:00:00                             1
#>  8 2016-07-29 00:00:00                             0
#>  9 2016-07-30 00:00:00                             0
#> 10 2016-07-31 00:00:00                             5
#> # … with 14 more rows