Datetime filtering for mts_monitor objects

Subsets a mts_monitor object by datetime. This function allows for sub-day filtering as opposed to monitor_filterDate() which always filters to day-boundaries.

Datetimes can be anything that is understood by MazamaCoreUtils::parseDatetime(). For non-POSIXct values, the recommended format is "YYYY-mm-dd HH:MM:SS".

Timezone determination precedence assumes that if you are passing in POSIXct values then you know what you are doing.

get timezone from startdate if it is POSIXct

use passed in timezone

get timezone from mts_monitor

monitor_filterDatetime(
  monitor = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  unit = "sec",
  ceilingStart = FALSE,
  ceilingEnd = FALSE
)

Arguments

monitor
mts_monitor object.

startdate
Desired start datetime (ISO 8601).

enddate
Desired end datetime (ISO 8601).

timezone
Olson timezone used to interpret dates.

unit
Units used to determine time at end-of-day.

ceilingStart
Logical instruction to apply ceiling_date to the startdate rather than floor_date

ceilingEnd
Logical instruction to apply ceiling_date to the enddate rather than floor_date

Value

A subset of the given mts_monitor object.

See also