Replace mts_monitor data with another value

Use an R expression to identify values for replacement.

The RR expression given in filter is used to identify elements in monitor$data that should be replaced. Typical usage would include

replacing negative values with 0

replacing unreasonably high values with NA

Expressions should use data for the left hand side of the comparison.

monitor_replaceValues(monitor = NULL, filter = NULL, value = NULL)

Arguments

monitor
mts_monitor object.

filter
RR expression used to identify values for replacement.

value
Numeric replacement value.

Examples

library(AirMonitor)

wa <- monitor_filterMeta(example_88101, stateCode == 'WA')
any(wa$data < 0, na.rm = TRUE)

#> [1] FALSE

wa_zero <- monitor_replaceValues(wa, data < 0, 0)
any(wa_zero$data < 0, na.rm = TRUE)

#> [1] FALSE