Convert monitor data as CSV

Converts the contents of the monitor object to CSV. By default, the output is a text string with "human readable" CSV that includes both meta and data. When saved as a file, this format is useful for point-and-click spreadsheet users who want to have everything on a single sheet.

To obtain a machine parseable CSV string for just the data, you can use includeMeta = FALSE. To obtain machine parseable metadata, use includeData = FALSE.

monitor_toCSV(monitor, includeMeta = TRUE, includeData = TRUE)

Arguments

monitor
mts_monitor object.

includeMeta
Logical specifying whether to include monitor$meta.

includeData
Logical specifying whether to include monitor$data.

Value

CSV formatted text.

Examples

library(AirMonitor)

data("Carmel_Valley")

Carmel_Valley <-
  Carmel_Valley %>%
  monitor_filterDate(20160802, 20160803)

monitor_toCSV(Carmel_Valley) %>% cat()

#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if `.name_repair` is omitted as of tibble 2.0.0.
#> Using compatibility `.name_repair`.

#> ##### Site metadata begins below here,NA,NA
#> parameter,blank,a9572a904a4ed46d_060530002_03
#> deviceDeploymentID,,a9572a904a4ed46d_060530002_03
#> deviceID,,060530002_03
#> deviceType,,NA
#> deviceDescription,,Met One BAM-1020 Mass Monitor w/VSCC - Beta Attenuation
#> deviceExtra,,NA
#> pollutant,,PM2.5
#> units,,UG/M3
#> dataIngestSource,,EPA_AQS
#> dataIngestURL,,https://aqs.epa.gov/aqsweb/airdata/download_files.html#Raw
#> dataIngestUnitID,,NA
#> dataIngestExtra,,NA
#> dataIngestDescription,,EPA AQS hourly data for parameterCode 88101
#> locationID,,a9572a904a4ed46d
#> locationName,,Carmel Valley
#> longitude,,-121.7333
#> latitude,,36.48187
#> elevation,,131
#> countryCode,,US
#> stateCode,,CA
#> countyName,,Monterey
#> timezone,,America/Los_Angeles
#> houseNumber,,NA
#> street,,NA
#> city,,Carmel Valley Village
#> zip,,93924
#> AQSID,,060530002
#> ##### Hourly data begins below here,NA,NA
#> UTC Time,Local Time,a9572a904a4ed46d_060530002_03
#> 2016-08-02 07:00:00 UTC,2016-08-02 00:00:00 PDT,26
#> 2016-08-02 08:00:00 UTC,2016-08-02 01:00:00 PDT,31
#> 2016-08-02 09:00:00 UTC,2016-08-02 02:00:00 PDT,38
#> 2016-08-02 10:00:00 UTC,2016-08-02 03:00:00 PDT,47
#> 2016-08-02 11:00:00 UTC,2016-08-02 04:00:00 PDT,76
#> 2016-08-02 12:00:00 UTC,2016-08-02 05:00:00 PDT,99
#> 2016-08-02 13:00:00 UTC,2016-08-02 06:00:00 PDT,126
#> 2016-08-02 14:00:00 UTC,2016-08-02 07:00:00 PDT,142
#> 2016-08-02 15:00:00 UTC,2016-08-02 08:00:00 PDT,156
#> 2016-08-02 16:00:00 UTC,2016-08-02 09:00:00 PDT,129
#> 2016-08-02 17:00:00 UTC,2016-08-02 10:00:00 PDT,126
#> 2016-08-02 18:00:00 UTC,2016-08-02 11:00:00 PDT,100
#> 2016-08-02 19:00:00 UTC,2016-08-02 12:00:00 PDT,90
#> 2016-08-02 20:00:00 UTC,2016-08-02 13:00:00 PDT,33
#> 2016-08-02 21:00:00 UTC,2016-08-02 14:00:00 PDT,16
#> 2016-08-02 22:00:00 UTC,2016-08-02 15:00:00 PDT,8
#> 2016-08-02 23:00:00 UTC,2016-08-02 16:00:00 PDT,9
#> 2016-08-03 00:00:00 UTC,2016-08-02 17:00:00 PDT,9
#> 2016-08-03 01:00:00 UTC,2016-08-02 18:00:00 PDT,9
#> 2016-08-03 02:00:00 UTC,2016-08-02 19:00:00 PDT,9
#> 2016-08-03 03:00:00 UTC,2016-08-02 20:00:00 PDT,6
#> 2016-08-03 04:00:00 UTC,2016-08-02 21:00:00 PDT,5
#> 2016-08-03 05:00:00 UTC,2016-08-02 22:00:00 PDT,6
#> 2016-08-03 06:00:00 UTC,2016-08-02 23:00:00 PDT,9

monitor_toCSV(Carmel_Valley, includeData = FALSE) %>% cat()

#> deviceDeploymentID,deviceID,deviceType,deviceDescription,deviceExtra,pollutant,units,dataIngestSource,dataIngestURL,dataIngestUnitID,dataIngestExtra,dataIngestDescription,locationID,locationName,longitude,latitude,elevation,countryCode,stateCode,countyName,timezone,houseNumber,street,city,zip,AQS_POC,AQS_GMT.Offset,AQS_Owning.Agency,AQS_Address,AQS_CBSA.Name,AQSID
#> a9572a904a4ed46d_060530002_03,060530002_03,NA,Met One BAM-1020 Mass Monitor w/VSCC - Beta Attenuation,NA,PM2.5,UG/M3,EPA_AQS,https://aqs.epa.gov/aqsweb/airdata/download_files.html#Raw,NA,NA,EPA AQS hourly data for parameterCode 88101,a9572a904a4ed46d,Carmel Valley,-121.73333,36.48187,131,US,CA,Monterey,America/Los_Angeles,NA,NA,Carmel Valley Village,93924,3,-8,Monterey Bay Unified APCD,35 Ford Road,"Salinas, CA",060530002

monitor_toCSV(Carmel_Valley, includeMeta = FALSE) %>% cat()

#> UTC Time,Local Time,a9572a904a4ed46d_060530002_03
#> 2016-08-02 07:00:00 UTC,2016-08-02 00:00:00 PDT,26
#> 2016-08-02 08:00:00 UTC,2016-08-02 01:00:00 PDT,31
#> 2016-08-02 09:00:00 UTC,2016-08-02 02:00:00 PDT,38
#> 2016-08-02 10:00:00 UTC,2016-08-02 03:00:00 PDT,47
#> 2016-08-02 11:00:00 UTC,2016-08-02 04:00:00 PDT,76
#> 2016-08-02 12:00:00 UTC,2016-08-02 05:00:00 PDT,99
#> 2016-08-02 13:00:00 UTC,2016-08-02 06:00:00 PDT,126
#> 2016-08-02 14:00:00 UTC,2016-08-02 07:00:00 PDT,142
#> 2016-08-02 15:00:00 UTC,2016-08-02 08:00:00 PDT,156
#> 2016-08-02 16:00:00 UTC,2016-08-02 09:00:00 PDT,129
#> 2016-08-02 17:00:00 UTC,2016-08-02 10:00:00 PDT,126
#> 2016-08-02 18:00:00 UTC,2016-08-02 11:00:00 PDT,100
#> 2016-08-02 19:00:00 UTC,2016-08-02 12:00:00 PDT,90
#> 2016-08-02 20:00:00 UTC,2016-08-02 13:00:00 PDT,33
#> 2016-08-02 21:00:00 UTC,2016-08-02 14:00:00 PDT,16
#> 2016-08-02 22:00:00 UTC,2016-08-02 15:00:00 PDT,8
#> 2016-08-02 23:00:00 UTC,2016-08-02 16:00:00 PDT,9
#> 2016-08-03 00:00:00 UTC,2016-08-02 17:00:00 PDT,9
#> 2016-08-03 01:00:00 UTC,2016-08-02 18:00:00 PDT,9
#> 2016-08-03 02:00:00 UTC,2016-08-02 19:00:00 PDT,9
#> 2016-08-03 03:00:00 UTC,2016-08-02 20:00:00 PDT,6
#> 2016-08-03 04:00:00 UTC,2016-08-02 21:00:00 PDT,5
#> 2016-08-03 05:00:00 UTC,2016-08-02 22:00:00 PDT,6
#> 2016-08-03 06:00:00 UTC,2016-08-02 23:00:00 PDT,9